feat: move thorchain LP /hooks to /queries/hooks
