+++ src/pages/ThorChainLP/AvailablePools.tsx
@@ -14,8 +14,8 @@ import { useAppSelector } from 'state/store'
 
 import { PoolIcon } from './components/PoolIcon'
 import { PoolsHeader } from './components/PoolsHeader'
-import type { ParsedPool } from './hooks/usePools'
-import { usePools } from './hooks/usePools'
+import type { ParsedPool } from './queries/hooks/usePools'
+import { usePools } from './queries/hooks/usePools'
 
 export const lendingRowGrid: GridProps['gridTemplateColumns'] = {
   base: 'minmax(150px, 1fr) repeat(1, minmax(40px, max-content))',
