+++ src/pages/ThorChainLP/components/AddLiquitity/AddLiquidityInput.tsx
@@ -38,7 +38,7 @@ import { isSome } from 'lib/utils'
 import { THOR_PRECISION } from 'lib/utils/thorchain/constants'
 import { estimateAddThorchainLiquidityPosition } from 'lib/utils/thorchain/lp'
 import { AsymSide, type ConfirmedQuote } from 'lib/utils/thorchain/lp/types'
-import { usePools } from 'pages/ThorChainLP/hooks/usePools'
+import { usePools } from 'pages/ThorChainLP/queries/hooks/usePools'
 import { selectAssetById, selectAssets, selectMarketDataById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
